Improve temporary error handling in package building process (#1030)
Improve temporary error handling in package building process (#1008)

* improve temporary error handling in package building process

This commit introduces a retry mechanism for various commands in the package building scripts to enhance resilience against transient failures. A new script, «run-with-retry.sh», is added to encapsulate the retry logic, which is then utilized in multiple scripts including «wazuh-dashboard.sh», «wazuh-dashboard-plugins.sh», and others.

Additionally, the Dockerfile and related scripts are updated to ensure that the retry logic is consistently applied across the package building process, improving the overall reliability of the build system.

* improve temporary error handling in package building process

Removed hardcoded retry parameters from the scripts and replaced them with environment variables «RETRY_MAX_ATTEMPTS» and «RETRY_DELAY_SECONDS». This change allows for more flexible configuration of retry behavior during package building, enhancing error handling and making the scripts more adaptable to different environments.

* improve temporary error handling in package building process

Refactor the error handling in the package building scripts to ensure that temporary directories are cleaned up properly. This includes updating the handling of the «tmp_dir», «applications_dir», and «base_dir» variables to enhance the robustness of the build process. The changes also ensure that the cleanup function is triggered on both INT and EXIT signals, preventing leftover files from previous builds and improving overall reliability.

Co-authored-by: Guido Modarelli <38738725+guidomodarelli@users.noreply.github.com>